requirements.txt
azure-ai-textanalytics==5.0.0
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.26
azure-core==1.12.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msrest==0.6.21
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
python-decouple==3.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
